Added liquibase migration of users_favorite_places

+++ service-api/src/main/java/greencity/dto/place/FavoritePlaceDto.java
@@ -10,5 +10,5 @@ import lombok.*;
 @EqualsAndHashCode
 public class FavoritePlaceDto {
     private String name;
-    private int placeId;
+    private long placeId;
 }
